fix index.js
