+++ js/index.js
@@ -1,154 +1,9 @@
-import Game_menu from "./game_menu.js";
-import Game_setting from "./game_setting.js";
-import Hituji_game from "./hituji_game.js";
-import Game_result from "./game_result.js";
-
-class Window_setting extends Phaser.Scene {
-  constructor() {
-    super({ key: "window", active: true });
-  }
-
-  create() {
-    setTimeout(() => {
-      // --- Window設定 ---
-      const graphics = this.add.graphics();
-
-      graphics
-        .fillStyle(0xffffff, 1)
-        .fillRoundedRect(352, 302, 320, 160, 5, 5).depth = 0;
-
-      // --- テキスト---
-      const screen_text = this.add.text(
-        388,
-        328,
-        "フルスクリーン表示しますか？",
-        {
-          fontSize: "18px",
-          fontFamily: "nomal",
-          fill: "#3c3c3c",
-        }
-      );
-
-      screen_text.setPadding(4).depth = 1;
-
-      const alert_text = this.add.text(
-        388,
-        421,
-        "※フルスクリーン表示にしないとゲームが開始されません",
-        {
-          fontSize: "10px",
-          fill: "#3c3c3c",
-        }
-      );
-
-      alert_text.setPadding(4).depth = 1;
-
-      // --- ボタン---
-
-      // --- ✖ボタン・イベント ---
-      const cross_button = this.add.text(652, 310, "✖", {
-        fontSize: "16px",
-        fill: "#707070",
-      });
-
-      cross_button.setInteractive().on(
-        "pointerdown",
-        () => {
-          this.scene.start("window");
-        },
-        this
-      );
-      cross_button.depth = 1;
-
-      // ---「いいえ」ボタン/テキスト---
-      const cancel_graphics = this.add.graphics();
-
-      cancel_graphics
-        .fillStyle(0x707070, 1)
-        .fillRoundedRect(373, 363, 137, 40, 5).depth = 1;
-
-      const cancel_text = this.add.text(422, 375, "いいえ", {
-        fontSize: "16px",
-        fill: "#ffffff",
-      });
-
-      cancel_text.depth = 2;
-      cancel_text.setInteractive().on(
-        "pointerdown",
-        () => {
-          this.cameras.main.fadeOut(500);
-          this.cameras.main.once("camerafadeoutcomplete", () => {
-            this.scene.start("window");
-          });
-        },
-        this
-      );
-
-      // 「はい」ボタン/テキスト
-      const dicision_graphics = this.add.graphics();
-
-      dicision_graphics
-        .fillStyle(0x32b65e, 1)
-        .fillRoundedRect(515, 363, 137, 40, 5).depth = 1;
-
-      const dicision_text = this.add.text(571, 375, "はい", {
-        fontSize: "16px",
-        fill: "#ffffff",
-      });
-
-      dicision_text.depth = 2;
-      dicision_text.setInteractive().on(
-        "pointerdown",
-        () => {
-          dicision_text.removeInteractive();
-          this.cameras.main.fadeOut(500);
-          this.cameras.main.once("camerafadeoutcomplete", () => {
-            this.scene.start("logo");
-          });
-        },
-        this
-      );
-    }, 3000);
-  }
-}
-
-class Open_logo extends Phaser.Scene {
-  constructor() {
-    super({ key: "logo", active: false });
-  }
-
-  preload() {
-    // ロゴマーク
-    this.load.image("logo", "../img/logo_touka.png");
-  }
-
-  create() {
-    // 画像表示
-
-    // logo画像
-    const logo_image = this.add.image(500, 400, "logo");
-    logo_image.depth = 1;
-
-    // fadein/out
-
-    this.cameras.main.fadeIn(2000);
-    this.cameras.main.once("camerafadeincomplete", (camera) => {
-      camera.fadeOut(2000);
-    });
-
-    this.cameras.main.once("camerafadeoutcomplete", () => {
-      this.scene.start("game_menu");
-    });
-
-    this.tweens.add({
-      targets: logo_image,
-      ease: "Sine.easeInOut",
-      yoyo: true,
-      repeat: -1,
-      duration: 3000,
-    });
-  }
-}
+import Game_menu from "./game_menu";
+import Game_setting from "./game_setting";
+import Hituji_game from "./hituji_game";
+import Game_result from "./game_result";
+import Window_setting from "./window_setting";
+import Open_logo from "./open_logo";
 
 const config = {
   type: Phaser.AUTO,
